OWRS-343 #time 4h Separated and structured hw flags
The hardware flags and a few other constants were previously stored at
the top of the pwmlib.c file. They have now been moved to the pwm_hw
source and header files.

The header contains the extern'd declarations and the source file the
definitions/initilisations.

A separate function that allows the pwm clock divider to be set was
added to the library, but it stills need to be commented documentd.

The library test file was also modified to output the duty percentage
over serial as a hexadecimal representation of a double (as UARTprintf
does not support floats).

The library was bulit and tested against the test file and is acting
as expected.

Comments and other documentation still need to be added.